Leitura labial/requirements_python313.txt
moviepy==1.0.3
google-generativeai==0.3.2
opencv-python==4.9.0.80
numpy==1.26.0
Pillow==10.0.0